pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "fastrr"
version = "0.1.0"
description = "A memory layer for AI applications: single-workspace, Git-versioned storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.5",
    "gitpython>=3.1.46",
    "ollama>=0.6.1",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "redis>=7.2.0",
    "redisvl>=0.3.9",
]

[tool.pytest.ini_options]
testpaths = ["tests", "tests_intg"]
pythonpath = ["src"]
markers = [
    "integration: mark test as integration (requires Ollama with configured model).",
]
